Struct ncollide_utils::data::hash_map::HashMap

pub struct HashMap<K, V, H> {
    // some fields omitted
}

Alternative implementation of HashMap.

It is different from std::hash_map::HashMap because:

the hash function can be personalized

the hash table is separate from the data. Thus, the vector of entries is tight (no holes due to sparse hashing).

Methods

impl<K, V, H: HashFun<K>> HashMap<K, V, H>

fn new(h: H) -> HashMap<K, V, H>

Creates a new hash map.

fn new_with_capacity(capacity: usize, h: H) -> HashMap<K, V, H>

Creates a new hash map with a given capacity.

fn elements<'r>(&'r self) -> &'r [Entry<K, V>]

The elements added to this hash map.

This is a simple, contiguous array.

fn elements_mut<'r>(&'r mut self) -> &'r mut [Entry<K, V>]

The elements added to this hash map.

This is a simple, contiguous array.

fn len(&self) -> usize

The number of elements contained by this hashmap.

fn is_empty(&self) -> bool

Whether or not this hashmap is empty.

fn clear(&mut self)

Removes everything from this hashmap.

impl<K: PartialEq + Clone, V, H: HashFun<K>> HashMap<K, V, H>

fn remove_elem_at(&mut self, at: usize) -> bool

Removes the element at the specified position of the element array.

If the index is greater than the table length, it returns false.

impl<K: PartialEq, V, H: HashFun<K>> HashMap<K, V, H>

fn get_and_remove(&mut self, key: &K) -> Option<Entry<K, V>>

Removes an element and returns its value if it existed.

fn find_or_insert_lazy<'a, F>(&'a mut self, key: K, value: F) -> Option<&'a mut V> where F: FnMut() -> Option<V>

Same as self.insert_or_replace(key, value, false) but with value a function which is called iff. the value does not exist yet. If the functions returns None, nothing is inserted.

fn insert_or_replace<'a>(&'a mut self, key: K, value: V, replace: bool) -> &'a mut V

Inserts or replace an element.

Arguments.

key - key of the element to add.

value - value to add.

replace - if true the new value will replace the existing value. If false, the old value is kept if it exists.

fn contains_key(&self, key: &K) -> bool

Checks whether this hashmap contains a specific key.

fn find<'a>(&'a self, key: &K) -> Option<&'a V>

Finds a reference to the element with a given key.

fn insert(&mut self, key: K, value: V) -> bool

Inserts an element on the hash map.

fn remove(&mut self, key: &K) -> bool

Remove an element from the hash map.

fn find_mut<'a>(&'a mut self, key: &K) -> Option<&'a mut V>

Gets a mutable reference to an element of the hashmap.