Struct ncollide_utils::data::uid_remap::UidRemap

pub struct UidRemap<O> { /* fields omitted */ }

A set of values having large usize key.

Methods

impl<O> UidRemap<O>

fn new(small_uids: bool) -> UidRemap<O>

Creates an empty UidRemap.

fn get_fast_key(&self, uid: usize) -> Option<FastKey>

Gets the fast key associated to the given key.

fn len(&self) -> usize

Return the number of elements in the map.

fn is_empty(&self) -> bool

Return true if the map contains no elements.

fn clear(&mut self)

Clears the map, removing all key-value pairs.

fn get(&self, key: usize) -> Option<&O>

Returns a reference to the value corresponding to the key.

fn get_fast(&self, key: &FastKey) -> Option<&O>

Returns a reference to the value corresponding to the fast key.

fn contains_key(&self, key: usize) -> bool

Returns true if the map contains a value for the specified key.

fn contains_fast_key(&self, key: &FastKey) -> bool

Returns true if the map contains a value for the specified fast key.

fn get_mut(&mut self, key: usize) -> Option<&mut O>

Returns a mutable reference to the value corresponding to the key.

fn get_fast_mut<'a>(&'a mut self, key: &FastKey) -> Option<&'a mut O>

Returns a mutable reference to the value corresponding to the fast key.

fn insert(&mut self, uid: usize, value: O) -> (FastKey, Option<O>)

Inserts a key-value pair to the map. If the key already had a value present in the map, that value and its fast key are returned. Otherwise, None is returned.

fn remove(&mut self, uid: usize) -> Option<(FastKey, O)>

Removes a key from the map, returning the value at the key if the key exists.

fn keys<'a>(&'a self) -> FastKeys<'a, O>

Returns an iterator visiting all keys.

fn values<'a>(&'a self) -> Values<'a, O>

Returns an iterator visiting all values. The iterator's element type is &'r O.

fn iter<'a>(&'a self) -> FastKeysAndValues<'a, O>

Returns an iterator visiting all key-value pairs.

fn iter_mut<'a>(&'a mut self) -> FastKeysAndValuesMut<'a, O>

Returns an iterator visiting all key-value pairs with mutable references to the values.

impl<O: Clone> UidRemap<O>

fn update<F: Fn(O, O) -> O>(&mut self, key: usize, newval: O, ff: F) -> bool

Updates a value in the map. If the key already exists in the map, modifies the value with ff taking oldval, newval. Otherwise, sets the value to newval. Returns true if the key did not already exist in the map.

fn update_with_key<F: Fn(&usize, O, O) -> O>(&mut self,
key: usize,
val: O,
ff: F)
-> bool

Updates a value in the map. If the key already exists in the map, modifies the value with ff taking key, oldval, newval. Otherwise, sets the value to newval. Returns true if the key did not already exist in the map.