Module ncollide_utils::data::vec_slice

Slicing on non-contiguous data.

Structs

VecSlice
A vector slice with a specific length and stride.

VecSliceMut
A mutable vector slice with a specific length and stride.